AnonymousIframe: check FencedFrame interactions (1/3)
Embeds a FencedFrame into an anonymous iframe. Check whether the
anonymous bit is propagated or not.

Currently:
- base: fenced-frame not enabled.
- fenced-frame-shadow-dom: window.anonymous = true
- fenced-frame-mparch: window.anonymous = false

I discussed with clamy@. We think the anonymous bit should not propagate.
This avoid difficult questions about mixing both FencedFrame and
AnonymousFrame together.

We think the security properties still hold without the anonymous bit,
as long as we require the FencedFrame to adhere to the COEP policy of
its embedder in exchange.

Bug:1287458
Change-Id: I33f0ae891023f87665b077227ba6a81166fb7e34